src/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain-core
langchain-openai
langchain-text-splitters
transformers
torch
numpy
scikit-learn
spacy
pydantic
